Move ShapeBox to Typescript plus refactoring
Also moved files into a subfolder
Also reduced dependency on ElementRegistry to ShapeBox only.
